linting, improve task completion logic
