+++ apps/web/src/types/file-upload.ts
@@ -1,0 +1,7 @@
+export const SUPPORTED_FILE_TYPES = [
+  "image/jpeg",
+  "image/png",
+  "image/gif",
+  "image/webp",
+  "application/pdf",
+];
