+++ apps/web/src/types/stream.ts
@@ -1,0 +1,5 @@
+export type StateType = {
+  messages: any[]; // Using any for now, can be refined later
+  ui?: any[];
+  targetRepository?: { owner: string; repo: string };
+};
